Fix wrong alignment bug
Never gonna reach that point (anytime soon), but just in case

+++ src/utils/align.h
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace EGL3::Utils {
-    template<int Alignment, class N>
+    template<uint64_t Alignment, class N>
     static constexpr N Align(N Value) {
         return (N)(((uint64_t)Value + Alignment - 1) & ~(Alignment - 1));
     }
